Update README.md
The Iris Tumor Detection Using CNN project utilizes a client-server architecture to provide an efficient and user-friendly solution for detecting tumors in iris images. The frontend, developed using HTML, CSS, and JavaScript, ensures an interactive and intuitive interface for users to upload images and view results. The backend, built with Django, handles data processing, user management, and seamless integration of the trained CNN model. The deep learning model, developed using TensorFlow/Keras, is designed to analyze iris images and classify them as normal or indicative of a tumor. Image preprocessing, including resizing and normalization, is performed using OpenCV to prepare the data for analysis. Visualization tools like Matplotlib and Seaborn are employed to display results and insights, enhancing the overall usability of the system. This combination of technologies ensures accurate tumor detection, scalability, and an excellent user experience.